move db package to installer
