+++ pkg/fleet/installer/db/db_test.go
